requirements.txt
# Minimal dependencies for mock application
flask==1.1.4
werkzeug==1.0.1
flask-cors==4.0.0
gunicorn==21.2.0
flask-pwa==0.1.0  # For Progressive Web App support
Pillow==10.1.0
markupsafe==2.0.1
jinja2==2.11.3

# Dependencies for full application with TensorFlow
tensorflow==2.15.0
tensorflow-hub==0.15.0
tensorflow-datasets==4.9.3
tensorflow-addons==0.22.0
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
# Removed tflite-model-maker due to compatibility issues
tensorflowjs==4.13.0

# Development tools
# ipython==8.17.2
# jupyter==1.0.0